requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
dnspython==2.7.0
email_validator==2.2.0
frozenlist==1.6.0
google-auth==2.39.0
google-genai==1.13.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
loguru==0.7.3
magic-filter==1.0.12
multidict==6.4.3
packaging==25.0
pluggy==1.5.0
propcache==0.3.1
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pyrogram==2.0.106
PySocks==1.7.1
pytest==8.3.5
python-dotenv==1.1.0
python-telegram-bot==22.0
requests==2.32.3
rsa==4.9.1
sniffio==1.3.1
soupsieve==2.7
Telethon==1.40.0
TgCrypto==1.2.5
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
websockets==15.0.1
yarl==1.20.0
